refactor footer and readytotalk into layout folder
